+++ CHANGELOG.rst
@@ -35,6 +35,8 @@ Added
   StackStorm role mappings. This means that the same role can now be granted via multiple RBAC
   mapping files.
   #3763
+* Add new Jinja filters ``from_json_string``, ``from_yaml_string``, and ``jsonpath_query``.
+  #3763
 * Add new "Inquiry" capability, which adds ability to "ask a question", usually in a workflow.
   Create a new runner type: "inquirer" to support this, as well as new API endpoints and
   client commands for interacting with Inquiries